Verify that all detected interface are really interfaces
Before this commit all files in /sys/class/net were naively interpreted
as links to interfaces, which is not always the case.

A notable exception is a file name "bonding_masters" which is present,
when bonded interfaces are configured on the system.

This patch introduces code which checks whether the targeted file
is a directory and then assumes this is a valid interface.